More cpplint fixes.

+++ include/ofp/yaml/yactiontype.h
@@ -22,7 +22,7 @@ struct ScalarTraits<ofp::ActionType> {
 
   static StringRef input(StringRef scalar, void *ctxt, ofp::ActionType &value) {
     if (!value.parse(scalar)) {
-      unsigned long long num;
+      uint64_t num;
       if (llvm::getAsUnsignedInteger(scalar, 0, num) == 0) {
         value.setNative(ofp::UInt16_narrow_cast(num));
         return "";
