+++ include/ofp/yaml/yoxmtype.h
@@ -24,7 +24,7 @@ struct ScalarTraits<ofp::OXMType> {
 
   static StringRef input(StringRef scalar, void *ctxt, ofp::OXMType &value) {
     if (!value.parse(scalar)) {
-      unsigned long long num;
+      uint64_t num;
       if (llvm::getAsUnsignedInteger(scalar, 0, num) == 0) {
         value.setOxmNative(ofp::UInt32_narrow_cast(num));
         return "";
